% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{alfalfaDefaultDF}
\alias{alfalfaDefaultDF}
\title{Alfalfa default data table}
\usage{
alfalfaDefaultDF()
}
\value{
Alfalfa default data
}
\description{
The alfalfa default data table holds pre-defined data if there is no user
input data.
}
