% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{basicClimateRating}
\alias{basicClimateRating}
\title{Basic climate rating}
\usage{
basicClimateRating(
  ratingTableArrayMC,
  ratingTableArrayTF,
  ppe,
  temperatureFactor,
  type
)
}
\arguments{
\item{ratingTableArrayMC}{Rating table lower and upper bounds for deduction for the moisture component.}

\item{ratingTableArrayTF}{Rating table lower and upper bounds for deduction for temperature factors.}

\item{ppe}{Precipitation minus potential evapotranspiration}

\item{temperatureFactor}{Input effective growing degree days or crop heat units for
the study site.}

\item{type}{If the crop uses effective growing degree days (EGDD) use EGDD else
if the crop uses crop heat units (CHU) use CHU.}
}
\value{
Deduction points for the basic climate rating.
}
\description{
The basic climate rating is a designed to return the point deduction
for the moisture component and temperature factors. The max deduction is
taken for between the moisture factor and the temperature factor.
}
