% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_index.R
\name{basicLandscapeRating}
\alias{basicLandscapeRating}
\title{Basic landscape rating}
\usage{
basicLandscapeRating(slopePercent, slopeLength)
}
\arguments{
\item{slopePercent}{Slope percentage.}

\item{slopeLength}{Slope length based on LS calculation.}
}
\value{
Deduction points for the basic landscape rating.
}
\description{
The basic landscape rating returns the point deduction for the percent
slope and landscape type.
}
