% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{basicOrganicRating}
\alias{basicOrganicRating}
\title{Basic organic rating (subsurface factors)}
\usage{
basicOrganicRating(subsurfaceBD, subsurfacepH, subsurfaceSalinity)
}
\arguments{
\item{subsurfaceBD}{The subsurface bulk density Mg/m^3. Default is at 60cm depth.}

\item{subsurfacepH}{The subsurface pH measured in saturated paste (dS/m). Default is at 60cm depth.}

\item{subsurfaceSalinity}{The subsurface salinity measured in saturated paste (dS/m). Default is at 60cm depth.}
}
\description{
The basic organic rating (subsurface factors) are considered as modifiers
of the surface (base) rating. As such, the basic organic rating is a percentage
reduction. The maximum depth is considered at 120cm. There are four factors that
are recognized in this category, structure (degree of decomposition) (B), depth of
deposit and kind of substrate (G), reaction (V) and salinity (N). Depth of
deposit and kind of substrate are currently not being used.
}
