% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral_soil_index.R
\name{basicSoilRating}
\alias{basicSoilRating}
\title{Basic soil rating (subsurface factors)}
\usage{
basicSoilRating(
  subsurfaceSiltPercent,
  subsurfaceClayPercent,
  subsurfaceBulkDensity,
  impedingLayerDepth,
  ppe,
  subsurfacepH,
  subsurfaceSalinity,
  subsurfaceSodicity
)
}
\arguments{
\item{subsurfaceSiltPercent}{Percentage of subsurface (depths 60-200cm) silt}

\item{subsurfaceClayPercent}{Percentage of subsurface (depths 60-200cm) clay}

\item{subsurfaceBulkDensity}{Subsurface bulk density}

\item{impedingLayerDepth}{Subsurface impeding layers depth (cm)}

\item{subsurfaceSalinity}{Subsurface salinity (depths 60-200cm)}

\item{subsurfaceSodicity}{Subsurface sodicity (depths 60-200cm)}
}
\value{
Percentage deduction for basic soil rating.
}
\description{
The basic soil rating determines a point deduction based on the subsurface
parameters of the soil. There are five parameters for subsurface factors. These
are impeding layer (structure and consistence) (D), contrasting texture,
reaction (pH) (V), salinity (EC) (N), sodicity (SAR) (Y). Currently, E, N, Y,
and O are not used. Future updates will include this additional information.
}
