% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{batchCropRaster}
\alias{batchCropRaster}
\title{Batch Crop Raster}
\usage{
batchCropRaster(inputFolder, inputExtent)
}
\arguments{
\item{inputFolder}{Input raster folder for batch processing}

\item{inputExtent}{Extent to crop. Input is a shape file or raster}
}
\value{
Cropped input file
}
\description{
Batch crop a folder of rasters and save to temporary location
}
