% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{batchMaskFolder}
\alias{batchMaskFolder}
\title{Batch Mask Folder}
\usage{
batchMaskFolder(inputFolder, exportFolder)
}
\arguments{
\item{inputFolder}{Input folder}

\item{exportFolder}{Export folder. Location where data is saved.}
}
\value{
Masked input file
}
\description{
Batch mask a folder of files.
** This tool is meant to be used to mask dead space around two images. This
tool helps with issues of miss-aligned input files by filling the extra space
in the larger of the two input files with with NA data.**
}
