% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{batchRunSRSMain}
\alias{batchRunSRSMain}
\title{Batch run SRS main}
\usage{
batchRunSRSMain(cropTypeList, cropArrays, inputFolderLocation, inputShapefile)
}
\arguments{
\item{cropTypeList}{The crop type.}

\item{cropArrays}{The arrays used for each specific crop variable.}

\item{inputFolderLocation}{Input folder location.}

\item{inputShapefile}{Input shapefile with folder location.}
}
\value{
Results of the SRS main.
}
\description{
This tools is a development tool used to run the SRS main function with entered
crop types.
}
