% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{canolaDefaultDF}
\alias{canolaDefaultDF}
\title{Canola default data table}
\usage{
canolaDefaultDF()
}
\value{
Canola default data
}
\description{
The Canola default data table holds pre-defined data if there is no user
input data.
}
