% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{chickpeasDefaultDF}
\alias{chickpeasDefaultDF}
\title{Chickpeas default data table}
\usage{
chickpeasDefaultDF()
}
\value{
chickpeas default data
}
\description{
The chickpeas default data table holds pre-defined data if there is no user
input data.
}
