% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_library.R
\name{chuComponent}
\alias{chuComponent}
\title{Crop heat unit component}
\usage{
chuComponent(inputArray, chu)
}
\arguments{
\item{inputArray}{An input array with the upper and lower bounds for the
deduction in each rating.}

\item{chu}{Crop heat unit}
}
\value{
deduction value
}
\description{
Calculate the point deduction based on the crop heat unit component
}
