% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{climateIndexMain}
\alias{climateIndexMain}
\title{Climate Index Main}
\usage{
climateIndexMain(
  ratingTableArrayMC,
  ratingTableArrayTF,
  ratingTableArrayESM,
  ratingTableArrayEFM,
  ppe,
  temperatureFactor,
  ppeSpring,
  ppeFall,
  type
)
}
\arguments{
\item{ratingTableArrayMC}{Rating table lower and upper bounds for deduction for the moisture component.}

\item{ratingTableArrayTF}{Rating table lower and upper bounds for deduction for temperature factors.}

\item{ratingTableArrayESM}{Rating table lower and upper bounds for deduction for early spring moisture.}

\item{ratingTableArrayEFM}{Rating table lower and upper bounds for deduction for excess fall moisture.}

\item{ppe}{Precipitation minus potential evapotranspiration}

\item{temperatureFactor}{Input effective growing degree days or crop heat
units for the study site.}

\item{ppeSpring}{Precipitation minus potential evapotranspiration for spring/April}

\item{ppeFall}{Precipitation minus potential evapotranspiration for fall/October}

\item{type}{If the crop uses effective growing degree days (EGDD) use EGDD else
if the crop uses crop heat units (CHU) use CHU.}
}
\value{
Deduction points for the basic climate rating.
}
\description{
The climate index main calls all required function and produces the rating
for climate over the study site.
}
