% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{climateModifyingFactors}
\alias{climateModifyingFactors}
\title{Climate modifying factors}
\usage{
climateModifyingFactors(
  ratingTableArrayESM,
  ratingTableArrayEFM,
  ppeSpring,
  ppeFall
)
}
\arguments{
\item{ratingTableArrayESM}{Rating table lower and upper bounds for deduction for ESM.}

\item{ratingTableArrayEFM}{Rating table lower and upper bounds for deduction for EFM.}

\item{ppeSpring}{Precipitation minus potential evapotranspiration for spring}

\item{ppeFall}{Precipitation minus potential evapotranspiration for fall}
}
\description{
The climate modifying factors is a designed to return the percentage deduction
for early spring moisture, excess fall moisture, and early fall frost. Max
deduction for each modifying factor is limited to 10% max modification.
}
