% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{climateRating}
\alias{climateRating}
\title{Climate rating}
\usage{
climateRating(moistureFactor, temperatureFactor, modifiers)
}
\arguments{
\item{modifiers}{Modifying factors.}

\item{basicClimate}{Basic climate rating calculated}
}
\value{
The climate rating.
}
\description{
The climate rating calculates the rating class for the climate index.
}
