% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\docType{class}
\name{ClinicalExperiment-class}
\alias{ClinicalExperiment-class}
\alias{numberOfFactors}
\alias{numberOfFactors,ClinicalExperiment-method}
\alias{factorNames}
\alias{factorNames,ClinicalExperiment-method}
\alias{factorLevelNames}
\alias{factorLevelNames,ClinicalExperiment-method}
\alias{numberOfTreatments}
\alias{numberOfTreatments,ClinicalExperiment-method}
\alias{treatmentNames}
\alias{treatmentNames,ClinicalExperiment-method}
\alias{show,ClinicalExperiment-method}
\title{ClinicalExperiment: Object representing Clinical Experiment}
\usage{
numberOfFactors(object)

\S4method{numberOfFactors}{ClinicalExperiment}(object)

factorNames(object)

\S4method{factorNames}{ClinicalExperiment}(object)

factorLevelNames(object)

\S4method{factorLevelNames}{ClinicalExperiment}(object)

numberOfTreatments(object)

\S4method{numberOfTreatments}{ClinicalExperiment}(object)

treatmentNames(object)

\S4method{treatmentNames}{ClinicalExperiment}(object)

\S4method{show}{ClinicalExperiment}(object)
}
\arguments{
\item{object}{the ClinicalExperiment object}
}
\description{
A class to encapsulate a clinical experiment with its
associated arms and factors. Objects can be created by calls of
the form \code{new("ClinicalExperiment", ...)}, although the
\code{ClinicalExperiment} constructor function is recommended.The arguments
include the number of factors in the experiment, the names of
the factors, the number of levels of each factor, the number of
treatments and the treatment names. Many, if not all, of the
arguments are optional, in which case a two-treatment,
two-factor experiment with two levels of each factor is
assumed.
}
\section{Slots}{

\describe{
\item{\code{number.of.factors}}{an integer specifying the number of factors, defaults to 2}

\item{\code{factor.names}}{a character vector specifying the character names, defaults to \code{F1} and \code{F2}}

\item{\code{number.of.factor.levels}}{A vector specifying the number of levels of each factor, defaults to \code{c(2,2)}}

\item{\code{factor.level.names}}{a list of character vectors specifying the factor level names, defaults to \code{F1-1}, \code{F1-2}, \code{F2-1}, \code{F2-2}, etc.}

\item{\code{number.of.treatments}}{the number of treatments, defaults to 2}

\item{\code{treatment.names}}{the names of the treatments; defaults to \code{c("Tr1", "Tr2")}}
}}

\examples{

showClass("ClinicalExperiment")
##
## Construct a Clinical Experiment with 3 factors, with levels 2, 2, 3
## respectively, and three treatments with default names for all.
##
expt <- ClinicalExperiment(number.of.factors = 3,
                           number.of.factor.levels = c(2, 2, 3),
                           number.of.treatments = 3)
print(expt)
}
\seealso{
ClinicalExperiment
}
