% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{ClinicalExperiment}
\alias{ClinicalExperiment}
\title{Constructor function for ClinicalExperiment class}
\usage{
ClinicalExperiment(number.of.factors = 2, factor.names,
  number.of.factor.levels, factor.level.names, number.of.treatments = 2,
  treatment.names)
}
\arguments{
\item{number.of.factors}{the number of factors, defaults to 2}

\item{factor.names}{names of factors defaults to \code{F1} and \code{F2}}

\item{number.of.factor.levels}{the number of levels of each factor, defaults to \code{c(2,2)}}

\item{factor.level.names}{a list of character vectors specifying the factor level names, defaults to \code{F1-1}, \code{F1-2}, \code{F2-1}, \code{F2-2}, etc.}

\item{number.of.treatments}{the number of treatments, defaults to 2}

\item{treatment.names}{the names of the treatments; defaults to \code{c("Tr1", "Tr2")}}
}
\value{
a ClinicalExperiment class object
}
\description{
Constructor function for ClinicalExperiment class
}
\examples{

expt <- ClinicalExperiment(number.of.factors = 3,
                           number.of.factor.levels = c(2, 2, 3),
                           number.of.treatments = 3)

}
