\name{computeImbalances-methods}
\docType{methods}
\alias{computeImbalances-methods}
\alias{computeImbalances,PocockSimonRandomizer,character-method}
\title{Methods for Function computeImbalances}
\description{
  This function computes the imbalances resulting from assigning each
  of the treatments in turn for a prospective subject coming in with the
  factors specified
}
\section{Methods}{
\describe{
\item{object = "PocockSimonRandomizer", factor.values = "character"}{
  Given a prospective subject with specified factor levels, compute the
  imbalances for each treatment from assigning the subject to each of
  the treatements}
}}
\keyword{methods}
