% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{cornDefaultDF}
\alias{cornDefaultDF}
\title{Corn default data table}
\usage{
cornDefaultDF()
}
\value{
Corn default data
}
\description{
The corn default data table holds pre-defined data if there is no user
input data.
}
