% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{dataPrep}
\alias{dataPrep}
\title{Data preparation tool}
\usage{
dataPrep(index, requiredDataArray, rasterStackFolder, shapefileAOI)
}
\arguments{
\item{index}{The index is the type of index the user is requesting. Options
are climate, soil, or landscape. Future versions will expand the indices.}

\item{shapefileAOI}{Area of interest for the SRS calculation. Shapefile format.}
}
\description{
This tools prepares input raster data for further processing. The result is
a raster stack that will be later used to calculate the rating tables
for climate, mineral, organic, and landscape indices.
}
