% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{deleteFolder}
\alias{deleteFolder}
\title{Delete folder}
\usage{
deleteFolder(inputFile)
}
\arguments{
\item{inputFile}{Input folder that will be deleted}
}
\description{
Delete the folder in the file path
}
