% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{drybeansDefaultDF}
\alias{drybeansDefaultDF}
\title{Dry beans default data table}
\usage{
drybeansDefaultDF()
}
\value{
dry beans default data
}
\description{
The dry beans default data table holds pre-defined data if there is no user
input data.
}
