% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_library.R
\name{effComponent}
\alias{effComponent}
\title{Early fall frost component}
\usage{
effComponent(inputArray, daysBeforeFrost)
}
\arguments{
\item{inputArray}{An input array with the upper and lower bounds for the
deduction in each rating.}

\item{daysBeforeFrost}{Number of days before average fall frost.}
}
\value{
deduction value
}
\description{
Calculate the percent deduction based on the early fall frost component
}
