% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{eraseTempData}
\alias{eraseTempData}
\title{Erase temp data}
\usage{
eraseTempData(file)
}
\arguments{
\item{file}{The file to be deleted.}
}
\value{
Confirmation of deletion
}
\description{
Erase temporary data
}
