% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{fababeansDefaultDF}
\alias{fababeansDefaultDF}
\title{Faba beans default data table}
\usage{
fababeansDefaultDF()
}
\value{
faba beans default data
}
\description{
The faba beans default data table holds pre-defined data if there is no user
input data.
}
