% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{FFP}
\alias{FFP}
\title{Full path string of file}
\usage{
FFP(inputFile)
}
\arguments{
\item{inputFile}{Input file that needs to return the full path}
}
\value{
String of full path of the file
}
\description{
Used to get the full path of a file.
}
