% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_index.R
\name{interimLandscapeRating}
\alias{interimLandscapeRating}
\title{Interim landscape rating}
\usage{
interimLandscapeRating(surfaceStoniness, coarseFragment, woodContent)
}
\arguments{
\item{surfaceStoniness}{Surface stoniness in annual removal (cubic m/ha)}

\item{coarseFragment}{Coarse fragment content as a percentage of volume.}

\item{woodContent}{Wood content as a percentage of volume.}
}
\value{
Deduction points for the interim landscape rating.
}
\description{
The interim landscape rating returns the point deduction as a percent deduction
from the basic landscape rating. This parameter is currently not being used.
}
