% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{interimOrganicRating}
\alias{interimOrganicRating}
\title{Interim organic rating (surface factors)}
\usage{
interimOrganicRating(surfaceBD, ppe, surfacepH, surfaceSalinity)
}
\arguments{
\item{surfaceBD}{Surface bulk density Mg/m^3}

\item{ppe}{Precipitation minus potential evapotranspiration}

\item{surfacepH}{Surface pH measured in saturated paste}

\item{surfaceSalinity}{Surface salinity measured in saturated paste (dS/m)}
}
\value{
Point deduction for interim organic rating (surface factors).
}
\description{
The interim organic rating returns the point deduction for the surface factors
of organic soils. The top 60cm of compacted peat is considered for the base
rating. Three factors are rated for their contribution to seed establishment,
crop growth and management. These are structure and consistence/degree of
decomposition (B) (fibre content), reaction (V) and nutrient status and
salinity (N).
}
