% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral_soil_index.R
\name{interimSoilRating}
\alias{interimSoilRating}
\title{Interim soil rating (surface factors)}
\usage{
interimSoilRating(
  surfaceSiltPercent,
  surfaceClayPercent,
  surfaceOC,
  depthOfTopSoil,
  surfacepH,
  surfaceSalinity,
  surfaceSodicity,
  depthOfPeat
)
}
\arguments{
\item{surfaceSiltPercent}{Percentage of surface (depths 0-60cm) silt}

\item{surfaceClayPercent}{Percentage of surface (depths 0-60cm) clay}

\item{surfaceOC}{Percentage of surface (depths 0-60cm) organic carbon}

\item{depthOfTopSoil}{Depth of top soil up to a max of 20cm}

\item{surfacepH}{Surface pH measured in saturated paste (depths 0-60cm)}

\item{surfaceSalinity}{Surface salinity (depths 0-60cm)}

\item{surfaceSodicity}{Surface sodicity (depths 0-60cm)}

\item{depthOfPeat}{Depth of organic (peaty) surface}
}
\value{
Deduction points for interim soil rating.
}
\description{
The interim soil rating determines a point deduction based on the surface
parameters of the soil. There are seven parameters for surface factors. These
are structure and consistence (D), organic matter context (colour) (F),
depth of top soil (E), reaction (pH) (V), salinity (dS/m) (N), sodicity (SAR) (Y),
and peat depth (cm) (O). Currently, E, N, Y, and O are not used. Future updates
will include this additional information.
}
