% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.R
\name{is.ClinicalExperiment}
\alias{is.ClinicalExperiment}
\title{Is an object a ClinicalExperiment object?}
\usage{
is.ClinicalExperiment(object)
}
\arguments{
\item{object}{an object}
}
\value{
TRUE if x is a ClinicalExperiment object, FALSE otherwise
}
\description{
Is an object a ClinicalExperiment object?
}
