% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_index.R
\name{landscapeIndexMain}
\alias{landscapeIndexMain}
\title{Landscape Index Main}
\usage{
landscapeIndexMain(
  slopePercent,
  slopeLength,
  surfaceStoniness,
  coarseFragment,
  woodContent
)
}
\value{
Landscape rating
}
\description{
The landscape index main calls all required function and produces the rating
for landscape over the study site.
}
