% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_index.R
\name{landscapeRating}
\alias{landscapeRating}
\title{Landscape rating}
\usage{
landscapeRating(basicLandscape, coarseFragmentModifications, otherModifiers)
}
\arguments{
\item{basicLandscape}{Basic landscape rating calculated}

\item{coarseFragmentModifications}{Coarse fragment modifications.}

\item{otherModifiers}{Other modifying factors such as pattern and flooding.}
}
\value{
The landscape rating.
}
\description{
The landscape rating calculates the rating class for the landscape index.
}
