% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{lentilsDefaultDF}
\alias{lentilsDefaultDF}
\title{Lentils default data table}
\usage{
lentilsDefaultDF()
}
\value{
Lentils default data
}
\description{
The Lentils default data table holds pre-defined data if there is no user
input data.
}
