% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{loadRaster}
\alias{loadRaster}
\title{Load Raster}
\usage{
loadRaster(inputRaster)
}
\arguments{
\item{inputRaster}{Location of input raster file}
}
\value{
Raster
}
\description{
Load raster into work space
}
