% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{loadShapefile}
\alias{loadShapefile}
\title{Shapefile (Vector file)}
\usage{
loadShapefile(shapefile)
}
\arguments{
\item{shapefile}{Location of input shapefile}
}
\value{
Shapefile
}
\description{
Load shapefile into work space
}
