% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_factor.R
\name{lsFactor}
\alias{lsFactor}
\title{LS factor}
\usage{
lsFactor(DEM)
}
\arguments{
\item{DEM}{Input digital elevation model that will be used to calculate the LS factor.}
}
\value{
The ls factor and the slope over the DEM.
}
\description{
The ls factor is the main function that calls the required functions to create
an ls layer. This code has been adopted and converted from the original work
completed by Amanda Moody, Central Washington University, Ellensburg, WA. 2019.
Although adopted from the source, the code has been changed significantly to increase
speed, remove redundent code and improve overall performance.
}
