% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_factor.R
\name{maxDHSArray}
\alias{maxDHSArray}
\title{Max downhill slope array}
\usage{
maxDHSArray(elevPixel, flowDir)
}
\arguments{
\item{elevPixel}{Elevation of the pixel}

\item{flowDir}{This is the flow direction.}
}
\value{
The max downhill slope for each cell
}
\description{
The max downhill slope array function
}
