% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{maxFunction}
\alias{maxFunction}
\title{Final table function}
\usage{
maxFunction(w, x, y, z)
}
\arguments{
\item{w, x, y, z}{The input climate, mineral soil, organic soil, and landscape
calculated values from each index.}
}
\value{
Returns a final rating from 1 to 7 with 1 being the best. Also returns
the index values for each respectable index. Climate, mineral soil, organic soil,
and landscape are the 2 to n values on each rating
}
\description{
This tools converts the values from each indices into a rating.
}
