% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral_soil_index.R
\name{mineralSoilIndexMain}
\alias{mineralSoilIndexMain}
\title{Mineral Soil Index Main}
\usage{
mineralSoilIndexMain(
  ppe,
  surfaceSiltPercent,
  surfaceClayPercent,
  subsurfaceSiltPercent,
  subsurfaceClayPercent,
  waterTableDepth,
  surfaceOC,
  depthOfTopSoil,
  surfacepH,
  surfaceSalinity,
  surfaceSodicity,
  depthOfPeat,
  subsurfaceBulkDensity,
  impedingLayerDepth,
  subsurfacepH,
  subsurfaceSalinity,
  subsurfaceSodicity
)
}
\value{
test
}
\description{
The mineral soil index main calls all required function and produces the rating
for mineral soil over the study site.
}
