% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral_soil_index.R
\name{mineralSoilMoistureDeduction}
\alias{mineralSoilMoistureDeduction}
\title{Mineral soil moisture deduction}
\usage{
mineralSoilMoistureDeduction(
  ppe,
  surfaceSiltPercent,
  surfaceClayPercent,
  subsurfaceSiltPercent,
  subsurfaceClayPercent,
  waterTableDepth
)
}
\arguments{
\item{ppe}{Precipitation minus potential evapotranspiration}

\item{surfaceSiltPercent}{Percentage of surface (depths 0-60cm) silt}

\item{surfaceClayPercent}{Percentage of surface (depths 0-60cm) clay}

\item{subsurfaceSiltPercent}{Percentage of subsurface (depths 61-200cm) silt}

\item{subsurfaceClayPercent}{Percentage of subsurface (depths 61-200cm) clay}

\item{waterTableDepth}{Water table depth (in cm)}
}
\value{
Deduction points for the moisture deduction.
}
\description{
The moisture deduction returns the point deduction for the available water
holding capacity "AWHC", the surface texture, subsurface texture, and water
table depth. The deduction is divided into two part, the first being the
texture deduction using the AWHC, surface and subsurface textures. The second
part takes a removes a percentage of the deduction based on the water table
depth.
}
