% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral_soil_index.R
\name{mineralSoilRating}
\alias{mineralSoilRating}
\title{Mineral Soil rating

The mineral soil rating calculates the rating class for the mineral soil index.}
\usage{
mineralSoilRating(moistureDeduct, surfaceFactors, subsurfaceFactors, drainage)
}
\arguments{
\item{moistureDeduct}{Basic climate rating calculated}

\item{surfaceFactors}{Modifying factors.}

\item{subsurfaceFactors}{Precipitation minus potential evapotranspiration for fall}
}
\value{
The climate rating.
}
\description{
Mineral Soil rating

The mineral soil rating calculates the rating class for the mineral soil index.
}
