% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_library.R
\name{moistureComponent}
\alias{moistureComponent}
\title{Moisture component}
\usage{
moistureComponent(inputArray, ppe)
}
\arguments{
\item{inputArray}{An input array with the upper and lower bounds for the
deduction in each rating.}

\item{ppe}{Precipitation minus potential evapotranspiration}
}
\value{
deduction value
}
\description{
Calculate the point deduction based on the moisture component
}
