% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{moistureFactorRating}
\alias{moistureFactorRating}
\title{Moisture factor rating}
\usage{
moistureFactorRating(ratingTableArrayMC, ppe)
}
\arguments{
\item{ratingTableArrayMC}{Rating table lower and upper bounds for deduction for the moisture component.}

\item{ppe}{Precipitation minus potential evapotranspiration}
}
\value{
Deduction points for the basic climate rating.
}
\description{
Moisture factor rating
}
