% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{organicBaseRating}
\alias{organicBaseRating}
\title{Organic base rating (Z)}
\usage{
organicBaseRating(egdd)
}
\arguments{
\item{egdd}{Effective growing degree days.}
}
\value{
Return the organic base rating
}
\description{
The organic base rating looks at the micro climates of the organic soils.
Organic soils are commonly colder than those of associated mineral soils.
Because of these differences in soil thermal properties and because organic
soils usually occupy low positions in the landscape, a temperature deduction
is considered for organic soils.
}
