% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{organicSCDF}
\alias{organicSCDF}
\title{Organic structure and consistence data frame}
\usage{
organicSCDF()
}
\value{
The organic structure and consistence data frame
}
\description{
The organic structure and consistence data frame helps determine the preparation
of a proper seedbed.
}
