% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{organicSoilIndexMain}
\alias{organicSoilIndexMain}
\title{Organic Soil Index Main}
\usage{
organicSoilIndexMain(
  egdd,
  ppe,
  surfaceBD,
  subsurfaceBD,
  depthToWaterTable,
  surfacepH,
  surfaceSalinity,
  subsurfacepH,
  subsurfaceSalinity
)
}
\value{

}
\description{
The organic soil index main calls all required function and produces the rating
for organic soil over the study site.
}
