% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{organicSoilMoistureDeduction}
\alias{organicSoilMoistureDeduction}
\title{Organic soil moisture deduction (M)}
\usage{
organicSoilMoistureDeduction(ppe, surfaceBD, subsurfaceBD, depthToWaterTable)
}
\arguments{
\item{ppe}{Precipitation minus potential evapotranspiration}

\item{surfaceBD}{Surface bulk density}

\item{subsurfaceBD}{Subsurface bulk density}

\item{depthToWaterTable}{Depth to water table in cm}
}
\value{
Organic soil moisture point deduction
}
\description{
The organic soil moisture deduction returns the point deduction for the water
supplying ability. The deduction is divided into two parts, the first being
the water holding capacity of the organic soil and the second being an adjustment
to part one, removing a percentage based on the water table depth.
}
