% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{peasDefaultDF}
\alias{peasDefaultDF}
\title{Peas default data table}
\usage{
peasDefaultDF()
}
\value{
peas default data
}
\description{
The peas default data table holds pre-defined data if there is no user
input data.
}
