% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{potatoDefaultDF}
\alias{potatoDefaultDF}
\title{Potato default data table}
\usage{
potatoDefaultDF()
}
\value{
Potato default data
}
\description{
The Potato default data table holds pre-defined data if there is no user
input data.
}
