\name{randomize-methods}
\docType{methods}
\alias{randomize-methods}
\alias{randomize,PocockSimonRandomizer,character,character-method}
\title{Methods for Function randomize}
\description{
  This method assigns a new treatment to a subject with a specified set
  of factors
}
\section{Methods}{
\describe{
\item{object = "PocockSimonRandomizer", subject.id = "character",
  factor.values = "character"}{ Given a subject id and a set of factor
  values, randomize to a given treatment}
}}
\keyword{methods}
