% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{ratingTable}
\alias{ratingTable}
\title{Rating table function}
\usage{
ratingTable(inputValue)
}
\arguments{
\item{inputValue}{The input values to be converted to a rating}
}
\value{
Returns a rating from 1 to 7 with 1 being the best
}
\description{
This tools converts the values from each indices into a rating.
}
