% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{reprojectFile}
\alias{reprojectFile}
\title{Reproject input data}
\usage{
reprojectFile(inputFile, projection)
}
\arguments{
\item{inputFile}{Input data}

\item{projection}{Projection to reproject to. Default is EPSG4326.}
}
\value{
Reprojected input data file
}
\description{
Reproject the input data to WGS 1984, EPSG4326
}
