% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{soilTexture}
\alias{soilTexture}
\title{Soil texture}
\usage{
soilTexture(siltPercent, clayPercent)
}
\arguments{
\item{siltPercent}{Percentage of silt in soil composition}

\item{clayPercent}{Percentage of clay in soil composition}
}
\value{
The texture of the soil profile based on percent silt clay
}
\description{
Determine soil texture based on percentage silt and clay percentages. Future
updates will reference soil lookup tables to best determine soil texture and
type.
}
