% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{soybeanDefaultDF}
\alias{soybeanDefaultDF}
\title{Soybean default data table}
\usage{
soybeanDefaultDF()
}
\value{
Soybean default data
}
\description{
The soybean default data table holds pre-defined data if there is no user
input data.
}
