\name{srs}
\docType{data}
\alias{srs}
\title{Sample srs dataset}
\description{Sample \code{srs} dataset.  To learn the format, you may do
  as follows.
  \preformatted{
    library(srs)
    data(srs)
    write.srs(srs)
  }
}
\usage{data(srs)}
\format{Contains records for two made-up students.}
\keyword{datasets}
