% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSampling.R
\name{SRSampling}
\alias{SRSampling}
\title{SRSampling: take one or multiple simple random samples}
\usage{
SRSampling(
  x,
  size,
  replace = FALSE,
  rep = 1,
  show.SRS = FALSE,
  verbose = TRUE,
  na.rm = FALSE,
  trim = 0,
  ...
)
}
\arguments{
\item{x}{the assumed population from which the SRS are chosen. It must be a vector with length > 1. NA in x can also be chosen.}

\item{size}{the sample size. the number of items to choose}

\item{replace}{whether sampling should be with replacement}

\item{rep}{the number of simulation runs. Default rep=1.}

\item{show.SRS}{whether to show SRS for each run of simulations}

\item{verbose}{whether to return verbose message}

\item{na.rm}{whether to remove NA when calculate summary statistics for SRS.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each end of x before the mean is computed. Values of trim outside that range are taken as the nearest endpoint}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a list a list containing the following components, conditional on the argument \code{rep}:
When \code{rep=1} (the default):
\itemize{
\item \code{sample.mean}: a numeric number giving the mean of the SRS
\item \code{sample.se}: a numeric number giving the SD of the SRS
\item \code{SRS}: a matrix containing the values in the SRS with \code{nrow=1} and \code{ncol=size}. Available only if \code{show.SRS=TRUE}
}
When \code{rep>1}:
\itemize{
\item \code{sample.mean}: a list containing the following components about sample means from multiple SRS:
\itemize{
\item \code{values}: a vector with a length of \code{rep} giving sample mean for each of the multiple SRS
\item \code{mean}: a numeric number giving the mean of \code{values}
\item \code{se}: a numeric number giving the mean of \code{values}
}
\item \code{sample.se}: a list containing the following components about sample sds from multiple SRS:
\itemize{
\item \code{values}: a vector with a length of \code{rep} giving sample sd for each of the multiple SRS
\item \code{mean}: a numeric number giving the mean of \code{values}
\item \code{se}: a numeric number giving the mean of \code{values}
}
\item \code{SRS}: a matrix containing the values in each of the multiple SRS with \code{nrow=rep} and \code{ncol=size}. Available only if \code{show.SRS=TRUE}.
}
}
\description{
SRSampling: take one or multiple simple random samples
}
\examples{
pop.1 <- rnorm(1000,0,10)
# take a SRS from x of size=10, show.SRS=TRUE
SRSampling(x=pop.1, size=10, show.SRS=TRUE)
# take 20 SRS from x of size=10, show.SRS=FALSE
SRSampling(x=pop.1, size=10, rep=20)

# size>length(pop.2)
pop.2 <- rnorm(10,0,10)
# SRSampling(x=pop.2, size=15, rep=20) # error

# non-integer size
SRSampling(x=pop.2, size=5.8) # warning

# x=numeric(0)
# SRSampling(x=numeric(0), size=1) # error

# x with NA
pop.3 <- c(1,2,3,4,NA)
SRSampling(x=pop.3, size=5, na.rm=TRUE)
SRSampling(x=pop.3, size=5)
}
