% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_main.R
\name{srsMain}
\alias{srsMain}
\title{SRS Main Function}
\usage{
srsMain(
  cropType,
  cropArrays,
  rasterStackFolder,
  shapefileAOI,
  indicesCalc,
  saveLocation
)
}
\value{

}
\description{
This function calls the data preparation tools and climate, mineral soil,
organic soil, and landscape indices.
}
