% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{SSSGDefaultDF}
\alias{SSSGDefaultDF}
\title{Spring seeded small grain (SSSG) default data table}
\usage{
SSSGDefaultDF()
}
\value{
Spring seeded small grain default data
}
\description{
The spring seeded small grain default data table holds pre-defined data if there is no user
input data.
}
