% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{subsurfaceAWHCDF}
\alias{subsurfaceAWHCDF}
\title{Subsurface available water holding capacity data table}
\usage{
subsurfaceAWHCDF()
}
\value{
The available water holding capacity data table.
}
\description{
The available water holding capacity (texture) aids in the determination of
the soils ability to retain and supply water to plants.
}
