% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{subsurfaceImpedingDF}
\alias{subsurfaceImpedingDF}
\title{Modifications for the impeding subsurface layers data frame}
\usage{
subsurfaceImpedingDF()
}
\value{
The modifications for the impeding subsurface layers data frame.
}
\description{
The modifications for the impeding subsurface layers data frame.
}
