% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_library.R
\name{subsurfaceSalinityComponent}
\alias{subsurfaceSalinityComponent}
\title{Subsurface salinity component}
\usage{
subsurfaceSalinityComponent(inputArray, ec)
}
\arguments{
\item{inputArray}{An input array with the upper and lower bounds for the
deduction in each rating.}

\item{ec}{Electro-conductivity}
}
\value{
deduction value
}
\description{
Calculate the point deduction based on the subsurface salinity component
}
