% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{subsurfaceSCDF}
\alias{subsurfaceSCDF}
\title{Subsurface structure and consistence data frame}
\usage{
subsurfaceSCDF()
}
\value{
The subsurface structure and consistence data frame.
}
\description{
The subsurface structure and consistence data frame.
}
