% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{subsurfaceWSADF}
\alias{subsurfaceWSADF}
\title{Water supplying ability data frame}
\usage{
subsurfaceWSADF()
}
\value{
The subsurface water supplying ability data frame
}
\description{
The subsurface water supplying ability helps determine the soils ability to retain and supply water to plants.
}
