% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{surfaceAndSubsurface}
\alias{surfaceAndSubsurface}
\title{Surface and subsurface average calculation}
\usage{
surfaceAndSubsurface(divideDepth, inputRaster)
}
\arguments{
\item{divideDepth}{Depth that delineates the difference between surface and
subsurface parameters. The depth is currently default to 60cm but the user can
specify an alternative depth. Example, any layer less than 60cm deep is
considered surface and any layer greater than 60cm is subsurface.}

\item{inputRaster}{The input raster that will be used to determine the average
surface and subsurface values.}
}
\value{
The average for the surface and subsurface along with average overall
}
\description{
Calculate the average surface value.
}
