% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceOrganicReactionDF}
\alias{surfaceOrganicReactionDF}
\title{Surface organic reaction (pH) data table}
\usage{
surfaceOrganicReactionDF()
}
\value{
Surface reaction (pH) data table
}
\description{
The surface organic reaction (pH) data table
}
