% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceOrganicSalinityDF}
\alias{surfaceOrganicSalinityDF}
\title{Surface organic salinity data table}
\usage{
surfaceOrganicSalinityDF()
}
\value{
Surface salinity (dS/m) data table
}
\description{
The surface organic salinity (dS/m) data table
}
