% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceReactionDF}
\alias{surfaceReactionDF}
\title{Surface mineral reaction (pH) data table}
\usage{
surfaceReactionDF()
}
\value{
Surface reaction (pH) data table
}
\description{
The surface reaction (pH) data table
}
