% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_library.R
\name{surfaceSalinityComponent}
\alias{surfaceSalinityComponent}
\title{Surface salinity component}
\usage{
surfaceSalinityComponent(inputArray, ec)
}
\arguments{
\item{inputArray}{An input array with the upper and lower bounds for the
deduction in each rating.}

\item{ec}{Electro-conductivity}
}
\value{
deduction value
}
\description{
Calculate the point deduction based on the surface salinity component
}
