% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceSalinityDF}
\alias{surfaceSalinityDF}
\title{Surface mineral salinity data table}
\usage{
surfaceSalinityDF()
}
\value{
Surface salinity (dS/m) data table
}
\description{
The surface salinity (dS/m) data table
}
