% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceSodicityDF}
\alias{surfaceSodicityDF}
\title{Surface sodicity data frame}
\usage{
surfaceSodicityDF()
}
\value{
surface sodicity (sodium adsorption ratio) data frame
}
\description{
The surface sodicity (sodium adsorption ratio) data frame
}
