% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{surfaceWSADF}
\alias{surfaceWSADF}
\title{Water supplying ability data frame}
\usage{
surfaceWSADF()
}
\value{
The water supplying ability data frame
}
\description{
The water supplying ability aid determines the soils ability to retain and supply water to plants.
}
