% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_index.R
\name{temperatureFactorRating}
\alias{temperatureFactorRating}
\title{Temperature Factor rating}
\usage{
temperatureFactorRating(ratingTableArrayTF, temperatureFactor, type)
}
\arguments{
\item{ratingTableArrayTF}{Rating table lower and upper bounds for deduction for temperature factors.}

\item{temperatureFactor}{Input effective growing degree days or crop heat units for
the study site.}

\item{type}{If the crop uses effective growing degree days (EGDD) use EGDD else
if the crop uses crop heat units (CHU) use CHU.}
}
\value{
Deduction points for the basic climate rating.
}
\description{
Temperature Factor rating
}
