\name{tr.assignments<-}
\alias{tr.assignments<-}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A function to set the treatment assigments slot }
\description{
  A function to set the treatment assigments slot
}
\usage{
tr.assignments(x)<- value
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{PocockSimonRandomizer} }
  \item{value}{ a data frame of factor values and treatment assignments }
}
\details{
  Some checks are made at this point for the value. However, this is not
  meant to be exhaustive and so care has to be taken. This is more often used in
  simulations and debugging than in ordinary situations.
}
\value{
}
\references{ }
\author{ Balasubramanian Narasimhan }
\note{ }
\seealso{ }
\examples{
###
### None
###
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ design }
\keyword{ distribution }% __ONLY ONE__ keyword per line
