% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{waterTableAdjustmentDF}
\alias{waterTableAdjustmentDF}
\title{Depth of water table adjustments data table}
\usage{
waterTableAdjustmentDF()
}
\value{
The water table adjustments
}
\description{
The depth of the water table effects the moisture supplying ability within
the root zone.
}
