% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wholenumber.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{is_wholenumber: check if a number is an integer}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric vector}

\item{tol}{the tolerance}
}
\value{
logical indicating whether x is integer
}
\description{
is_wholenumber: check if a number is an integer
}
\examples{
#is_wholenumber(1.5)
#is_wholenumber(1:5)
#is_wholenumber(c(1,2.5,NA))
}
