% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_index.R
\name{landscapeIndexMain}
\alias{landscapeIndexMain}
\title{Landscape Index Main}
\usage{
landscapeIndexMain(
  slopePercent,
  slopeLength,
  surfaceStoniness,
  coarseFragment,
  woodContent
)
}
\arguments{
\item{slopePercent}{Slope percentage.}

\item{slopeLength}{Slope length based on LS calculation.}

\item{surfaceStoniness}{Surface stoniness in annual removal (cubic m/ha)}

\item{coarseFragment}{Coarse fragment content as a percentage of volume.}

\item{woodContent}{Wood content as a percentage of volume.}
}
\value{
Landscape rating
}
\description{
The landscape index main calls all required function and produces the rating
for landscape over the study site.
}
