% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_soil_index.R
\name{organicSoilRating}
\alias{organicSoilRating}
\title{Organic Soil rating

The organic soil rating calculates the rating class for the organic soil index.}
\usage{
organicSoilRating(
  soilTemp,
  moistureDeduct,
  surfaceFactors,
  subsurfaceFactors,
  drainage
)
}
\arguments{
\item{basicClimate}{Basic climate rating calculated}

\item{modifiers}{Modifying factors.}

\item{ppeFall}{Precipitation minus potential evapotranspiration for fall}
}
\value{
The climate rating.
}
\description{

}
\details{

}
