% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{writePermData}
\alias{writePermData}
\title{Write permanent data}
\usage{
writePermData(inputData, exportLocation, exportName, exportType)
}
\arguments{
\item{inputData}{Input data that needs to be saved}
}
\value{
Saves the input data to a folder location
}
\description{
Write permanent data. This tool will write any data that is currently in
the same folder as the input data.
}
