% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operational_tools.R
\name{writeTempData}
\alias{writeTempData}
\title{Write temp data}
\usage{
writeTempData(inputData, tempFolder, tempName, exportType)
}
\arguments{
\item{inputData}{Input temp data that needs to be save temporarily}

\item{tempFolder}{Location where temp data will be saved}

\item{tempName}{Temp name for the files}

\item{exportType}{Export type of the file. Options are .shp or .grd/.gri}
}
\value{
Saves the input data to a temporary location
}
\description{
Write temporary data to temp folder for further processing
}
